Update passport.js
Removed the native call for pg - I'm honestly not sure what this does other than I know it requires a PostgreSQL install on the server as well as C++ redistributables (and that I spent 2 hours figuring out why it wasn't working on my dev machine...). As far as I know this isn't something that is do-able in Azure, so for now we will shortcut this and just load pg.